remove unnecessary int

+++ balanced_random/split/balanced_random_split.py
@@ -22,8 +22,8 @@ p.setup(timestep=0.1)
 p.set_number_of_neurons_per_core(p.IF_curr_exp, 64)
 p.set_number_of_neurons_per_core(p.SpikeSourcePoisson, 64)
 n_neurons = 500
-n_exc = int(round(n_neurons * 0.8))
-n_inh = int(round(n_neurons * 0.2))
+n_exc = round(n_neurons * 0.8)
+n_inh = round(n_neurons * 0.2)
 weight_exc = 0.1
 weight_inh = -5.0 * weight_exc
 weight_input = 0.001
